pdf.php: Annihilate hacked in RTF support
This libary was originally intended to support multiple document
formats, e.g. PDF, RTF (then MSWord, OpenDocument, etc.) however the
"best" libaries for RTF we could find have an API which is very
different to FPDF, so implementing RTF support would require
re-writing FPDF in terms of the RTF library then figuring out some
way to "filter" the positioning required to produce nice PDFs into a
format that works for RTFs. This would be a herculean task and is
well beyond the scope of this project.

Signed-off-by: Julian Calaby <julian.calaby@gmail.com>